requirements_flower.txt
# Flower Federated Learning Framework Requirements
# Install with: pip install -r requirements_flower.txt

# Core Flower framework
flwr[simulation]>=1.5.0

# Additional dependencies for enhanced functionality
numpy>=1.21.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
accelerate>=0.20.0
tqdm>=4.64.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
PyYAML>=6.0
datasets>=2.0.0
mmengine>=0.7.0

# Optional: For better performance
# numba>=0.56.0  # JIT compilation
# psutil>=5.9.0  # System monitoring
